♻️ Add Close button component and style right
