+++ ui/core/index.ts
@@ -3,15 +3,16 @@ export * from './Button'
 export * from './Callout'
 export * from './Card'
 export * from './Checkbox'
+export * from './CloseButton'
 export * from './Divider'
 export * from './Icon'
 export * from './IconButton'
 export * from './Input'
+export * from './List'
 export * from './Row'
 export * from './SRText'
 export * from './Table'
 export * from './Text'
-export * from './List'
 
 export interface As {
   as?: React.ElementType
